feat: add test for loading state
