+++ src/recommendations/tests/constants.js
@@ -1,0 +1,79 @@
+const mockedResponse = [
+  {
+    title: 'How to Learn Online 1',
+    marketingUrl: 'https://test-recommendations.com/course/how-to-learn-online-1',
+    cardImageUrl: 'https://test-recommendations.com/image/how-to-learn-online-1.png',
+    activeRunKey: 'course-v1:test+testX+2018',
+    owners: [
+      {
+        key: 'firstOwnerX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-1.png',
+        name: 'first owner',
+      },
+      {
+        key: 'secondOwnerX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-1.png',
+        name: 'second owner',
+      },
+    ],
+    objectId: 'course-how-to-learn-online-key-1',
+  },
+  {
+    title: 'How to Learn Online 2',
+    marketingUrl: 'https://test-recommendations.com/course/how-to-learn-online-2',
+    cardImageUrl: 'https://test-recommendations.com/image/how-to-learn-online-2.png',
+    activeRunKey: 'course-v1:test+testX+2019',
+    owners: [
+      {
+        key: 'testX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-2.png',
+        name: 'test',
+      },
+    ],
+    objectId: 'course-how-to-learn-online-key-2',
+  },
+  {
+    title: 'How to Learn Online 3',
+    marketingUrl: 'https://test-recommendations.com/course/how-to-learn-online-3',
+    cardImageUrl: 'https://test-recommendations.com/image/how-to-learn-online-3.png',
+    activeRunKey: 'course-v1:test+testX+2020',
+    owners: [
+      {
+        key: 'testX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-3.png',
+        name: 'test',
+      },
+    ],
+    objectId: 'course-how-to-learn-online-key-3',
+  },
+  {
+    title: 'How to Learn Online 4',
+    marketingUrl: 'https://test-recommendations.com/course/how-to-learn-online-4',
+    cardImageUrl: 'https://test-recommendations.com/image/how-to-learn-online-4.png',
+    activeRunKey: 'course-v1:test+testX+2021',
+    owners: [
+      {
+        key: 'testX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-4.png',
+        name: 'test',
+      },
+    ],
+    objectId: 'course-how-to-learn-online-key-4',
+  },
+  {
+    title: 'How to Learn Online 5',
+    marketingUrl: 'https://test-recommendations.com/course/how-to-learn-online-5',
+    cardImageUrl: 'https://test-recommendations.com/image/how-to-learn-online-5.png',
+    activeRunKey: 'course-v1:test+testX+2022',
+    owners: [
+      {
+        key: 'testX',
+        logoImageUrl: 'https://test-recommendations.com/logos/how-to-learn-online-5.png',
+        name: 'test',
+      },
+    ],
+    objectId: 'course-how-to-learn-online-key-5',
+  },
+];
+
+export default mockedResponse;
